Displays subscribed and unsubscribed

+++ telecom/src/main/java/com/googleit/telecom/dao/ServiceDAO.java
@@ -8,7 +8,7 @@ import java.util.List;
  * Created by Joon on 2/10/2015.
  */
 public interface ServiceDAO {
-    public List<Service> getUnsubscribedService(int user_id);
-    public List<Service> getSubscribedService(int user_id);
+    public List<Service> getUnsubscribedService(long user_id);
+    public List<Service> getSubscribedService(long user_id);
 
 }
